ci: update test trigger from pull_request to push
We previously set it to pull_request with the assumption that people working locally and then when pull_request. This will create a new env and the reviewer can use it. However most newbies aren't devops afficionados so be able when they create a new branch, deploying the infra would just be pushing the branch and then they can work. In addition, this also allow to check the deployed vesion befor merging.